Fix barlines scrolling at different speeds in legacy skins
